perf: replace sara with jwt decoder
