+++ .env
@@ -2,7 +2,6 @@ VUE_APP_SELF_HOST=https://web-tech.tw/openchat
 VUE_APP_OCJI_HOST=https://web-tech.tw/recv/openchat
 
 VUE_APP_SARA_INTE_HOST=https://web-tech.tw/sara
-VUE_APP_SARA_RECV_HOST=https://web-tech.tw/recv/sara
 VUE_APP_SARA_TOKEN_NAME=unified_token
 
 VUE_APP_TURNSTILE_SITE_KEY="0x4AAAAAAAr6LY1hBUbTBkVA"
