+++ package.json
@@ -12,6 +12,7 @@
     "autoprefixer": "^10",
     "core-js": "^3.8.3",
     "dayjs": "^1.11.12",
+    "jwt-decode": "^4.0.0",
     "nanoid": "^5.0.7",
     "postcss": "^8",
     "tailwindcss": "^3",
