+++ src/main.js
@@ -1,6 +1,6 @@
 import Vue from 'vue'
-import './plugins/sara'
 import './plugins/axios'
+import './plugins/profile'
 import App from './App.vue'
 import router from './router'
 import './assets/tailwind.css'
