+++ src/views/AdminJoinView.vue
@@ -176,7 +176,7 @@ export default {
     },
   },
   async created() {
-    this.profile = await this.$profile();
+    this.profile = this.$profile();
     if (!this.profile) {
       const refer = `${process.env.VUE_APP_SELF_HOST}/#/admin/join`;
       const url = `${process.env.VUE_APP_SARA_INTE_HOST}/?refer=${encodeURIComponent(refer)}`;
