+++ src/views/AdminRoomView.vue
@@ -62,7 +62,7 @@ export default {
     },
   },
   async created() {
-    this.profile = await this.$profile();
+    this.profile = this.$profile();
     if (!this.profile) {
       const refer = `${process.env.VUE_APP_SELF_HOST}/#/admin/room`;
       const url = `${process.env.VUE_APP_SARA_INTE_HOST}/?refer=${encodeURIComponent(refer)}`;
